requirements.txt
fastapi
uvicorn
langchain
openai
langchain_openai
langsmith
python-dotenv
